+++ apps/api/src/app/endpoints/sitemap/sitemap.service.ts
@@ -17,6 +17,121 @@ export class SitemapService {
     private readonly i18nService: I18nService
   ) {}
 
+  public getBlogPosts({ currentDate }: { currentDate: string }) {
+    const rootUrl = this.configurationService.get('ROOT_URL');
+
+    return [
+      {
+        languageCode: 'de',
+        routerLink: ['2021', '07', 'hallo-ghostfolio']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2021', '07', 'hello-ghostfolio']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '01', 'ghostfolio-first-months-in-open-source']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '07', 'ghostfolio-meets-internet-identity']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '07', 'how-do-i-get-my-finances-in-order']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '08', '500-stars-on-github']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '10', 'hacktoberfest-2022']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2022', '11', 'black-friday-2022']
+      },
+      {
+        languageCode: 'en',
+        routerLink: [
+          '2022',
+          '12',
+          'the-importance-of-tracking-your-personal-finances'
+        ]
+      },
+      {
+        languageCode: 'de',
+        routerLink: ['2023', '01', 'ghostfolio-auf-sackgeld-vorgestellt']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '02', 'ghostfolio-meets-umbrel']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '03', 'ghostfolio-reaches-1000-stars-on-github']
+      },
+      {
+        languageCode: 'en',
+        routerLink: [
+          '2023',
+          '05',
+          'unlock-your-financial-potential-with-ghostfolio'
+        ]
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '07', 'exploring-the-path-to-fire']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '08', 'ghostfolio-joins-oss-friends']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '09', 'ghostfolio-2']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '09', 'hacktoberfest-2023']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '11', 'black-week-2023']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2023', '11', 'hacktoberfest-2023-debriefing']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2024', '09', 'hacktoberfest-2024']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2024', '11', 'black-weeks-2024']
+      },
+      {
+        languageCode: 'en',
+        routerLink: ['2025', '09', 'hacktoberfest-2025']
+      }
+    ]
+      .map(({ languageCode, routerLink }) => {
+        return this.createRouteSitemapUrl({
+          currentDate,
+          languageCode,
+          rootUrl,
+          route: {
+            routerLink: ['blog', ...routerLink],
+            path: undefined
+          }
+        });
+      })
+      .join('\n');
+  }
+
   public getPersonalFinanceTools({ currentDate }: { currentDate: string }) {
     const rootUrl = this.configurationService.get('ROOT_URL');
 
@@ -43,20 +158,21 @@ export class SitemapService {
       });
 
       return personalFinanceTools.map(({ alias, key }) => {
-        const location = [
-          rootUrl,
-          languageCode,
+        const routerLink = [
           resourcesPath,
           personalFinanceToolsPath,
           `${productPath}-${alias ?? key}`
-        ].join('/');
+        ];
 
-        return [
-          '  <url>',
-          `    <loc>${location}</loc>`,
-          `    <lastmod>${currentDate}T00:00:00+00:00</lastmod>`,
-          '  </url>'
-        ].join('\n');
+        return this.createRouteSitemapUrl({
+          currentDate,
+          languageCode,
+          rootUrl,
+          route: {
+            routerLink,
+            path: undefined
+          }
+        });
       });
     }).join('\n');
   }