+++ apps/client/src/app/pages/blog/2025/09/hacktoberfest-2025/hacktoberfest-2025-page.component.ts
@@ -1,0 +1,17 @@
+import { publicRoutes } from '@ghostfolio/common/routes/routes';
+
+import { Component } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { RouterModule } from '@angular/router';
+
+@Component({
+  host: { class: 'page' },
+  imports: [MatButtonModule, RouterModule],
+  selector: 'gf-hacktoberfest-2025-page',
+  templateUrl: './hacktoberfest-2025-page.html'
+})
+export class Hacktoberfest2025PageComponent {
+  public routerLinkAbout = publicRoutes.about.routerLink;
+  public routerLinkBlog = publicRoutes.blog.routerLink;
+  public routerLinkOpenStartup = publicRoutes.openStartup.routerLink;
+}