+++ apps/client/src/app/pages/portfolio/activities/create-or-update-activity-dialog/types/activity-type.type.ts
@@ -1,0 +1,8 @@
+import { Type } from '@prisma/client';
+
+export const ActivityType = {
+  ...Type,
+  VALUABLE: 'VALUABLE'
+} as const;
+
+export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType];