+++ libs/ui/src/lib/i18n.ts
@@ -38,9 +38,9 @@ const locales = {
   DIVIDEND: $localize`Dividend`,
   FEE: $localize`Fee`,
   INTEREST: $localize`Interest`,
-  ITEM: $localize`Valuable`,
   LIABILITY: $localize`Liability`,
   SELL: $localize`Sell`,
+  VALUABLE: $localize`Valuable`,
 
   // AssetClass (enum)
   ALTERNATIVE_INVESTMENT: $localize`Alternative Investment`,